Added CORS, UUID, Deleting Temp Files, and Clean up Code
I first enabled CORS for the UI in order for the UI to make a request to the server without being blocked by CORS. I then changed some variables to use folder_name instead of "temp". I then stored the files in a new folder per request with the folder name being a UUID. After the results are gathered, delete all the temp files and folder that were created. Finally added the ability to just run "python server.py" instead of having to do "uvicorn server:app".

Note: Hot reloading does not work on windows but does on other platforms as the event loop is changed on windows platforms which doesn't allow asyncio.create_subprocess_exec to behave properly.